requirements.txt
flask==2.3.3
redis==5.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
redis==6.4.0